% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unsf}
\alias{unsf}
\title{Convert sf objects to normal data.frames with longitude and latitude columns}
\usage{
unsf(x, coord_names = c("lon", "lat"))
}
\arguments{
\item{x}{an \code{sf::sf} object}

\item{coord_names}{\code{character} vector of length \code{2}. The names of the new
that should contain the coordinates from the geometry column of \code{x}.}
}
\value{
a \code{data.frame}
}
\description{
Convert sf objects to normal data.frames with longitude and latitude columns
}
